fake volume メーターの位置を変更する
